+++ examples/test_heading_II.jl
@@ -163,7 +163,7 @@ function calc_clock_angle(turn_angle; x = 100.0, z = 0.0, r = 20.0, upwind_dir =
     el, az_east = calc_elevation_azimuth(turn_angle; x = x, z = z, r = r)
     az_north = wrap2pi(-π / 2 - az_east)
     az_wind = azn2azw(az_north; upwind_dir = upwind_dir)
-    KiteUtils.calc_clock_angle(orientation, el, az_wind; respos = false)
+    KiteUtils.calc_heading(orientation, el, az_wind; respos = false)
 end
 
 # Compute data for multiple θ values