+++ src/_log.jl
@@ -82,9 +82,6 @@ function log!(logger::Logger, state::SysState)
     logger.set_torque_vec[i] .= state.set_torque
     logger.set_speed_vec[i] .= state.set_speed
     logger.set_force_vec[i] .= state.set_force
-    logger.roll_vec[i] = state.roll
-    logger.pitch_vec[i] = state.pitch
-    logger.yaw_vec[i] = state.yaw
     logger.var_01_vec[i] = state.var_01
     logger.var_02_vec[i] = state.var_02
     logger.var_03_vec[i] = state.var_03